+++ src/Widget/WeatherWidget.php
@@ -6,8 +6,10 @@ namespace Bolt\Widget;
 
 use Bolt\Snippet\Target;
 
-class WeatherWidget extends BaseWidget implements TwigAware
+class WeatherWidget extends BaseWidget
 {
+    use TwigAware;
+
     protected $name = 'Weather Widget';
     protected $type = 'widget';
     protected $target = Target::WIDGET_BACK_DASHBOARD_ASIDE_TOP;
